Function refactoring and Addition of names and teams data. Dropped live data via intervals
